Improve SwitchToNotepadPlusPlusWindow
- Check for the foreground window's root owner, and if it's Notepad++ do
  not attempt a switch.

.. because it's possible Notepad++ may own a top-level enabled popup
that's the foreground window (eg "File xxx doesn't exist, create it?")
and which has disabled the main window. In such a case if Notepad++ is
not minimized then the popup is already the foreground and visible.

- In the steal focus hack explicitly call SW_RESTORE if
  SwitchToThisWindow can't restore the window.

.. because it's possible that a disabled main window may not restore
even if there are enabled popups.

- Try harder to set the foreground window to remedy a race condition.

.. because it's possible that if a previously existing mono-instance
of Notepad++ has a disabled main window then the just-started Notepad++
instance will try to set that as the foreground window (incorrectly in
my view) and terminate, at the same time this program attempts to set
the Notepad++ enabled popup as the foreground window.

- Add some programming notes in 'npp NOTES.txt'.