Wait longer for Notepad++ to load before switching to it
- Switch to the most recent Notepad++ window only after it stays the
  same for 1000ms, so that Notepad++ has a chance to load and put itself
  in the foreground on its own during that time.

This change also makes it so a Notepad++ instance that exists prior to
npp launching Notepad++ is less likely to be selected by mistake when
the just-launched instance is likely to stay present as a separate
instance (ie multi-instance) or have a popup for a mono-instance. In
that case it's preferable to select the just-launched window.