Fix bug that may occur when the window verification delay is 0
- Do not continue to the next iteration of the switch-to loop when the
  most recent main Notepad++ hwnd has changed, instead depend on the
  subsequent verification delay check to make that decision.

Prior to this change if the switch-to loop verification delay for hwnd
was set 0 and the max loop wait time was set to 0 then it's possible
starting another iteration would not complete due to various timing
checks at the beginning of the loop. In other words, it's possible there
would never be an opportunity to switch to the window in such a case.